+++ etc/sql/sql.sql
@@ -158,6 +158,7 @@ CREATE TABLE files (
     owner_id STRING,
     storage_type STRING,
     content_type STRING,
+    info OBJECT(IGNORED),
 
     INDEX original_name_ft
       USING FULLTEXT(original_name)