+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes For IRIS Service
 unreleased
 ==========
 
+ - added dimension and image proxy base url in files
+   DEPLOY: add column info in files table
  - added town property user relation in swagger spec
  - added file service (upload and download)
    DEPLOY: